requirements.txt
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
idna==3.7
pycparser==2.22
PyJWT==2.8.0
requests==2.32.3
urllib3==2.2.2
